+++ modules/Grades/ReportCardGrades.php
@@ -295,10 +295,12 @@ function _makeGradesInput( $value, $name )
 	}
 	elseif ( $name === 'BREAK_OFF'
 		&& $teacher_id
-		&& $gradebook_config[ UserCoursePeriod() . '-' . $THIS_RET['ID'] ][1]['VALUE'] != '' )
+		&& isset( $gradebook_config[ UserCoursePeriod() . '-' . $THIS_RET['ID'] ] )
+		&& $gradebook_config[ UserCoursePeriod() . '-' . $THIS_RET['ID'] ] != '' )
 	{
+		// Breakoff configured by Teacher.
 		return '<span style="color:blue">' .
-			$gradebook_config[ UserCoursePeriod() . '-' . $THIS_RET['ID'] ][1]['VALUE'] . '</span>';
+			$gradebook_config[ UserCoursePeriod() . '-' . $THIS_RET['ID'] ] . '%</span>';
 	}
 	else
 	{